fix ONB (Qt signal-slot connection on service msg)

+++ src/core/application.h
@@ -164,7 +164,7 @@ public:
 
     /*! Возвращает время в миллисекундах с момента запуска.
     */
-    inline unsigned long timestamp() const {return mTimestamp;}
+    inline uint32_t timestamp() const {return mTimestamp;}
 
     static void delay(int ms);
 };
